build(Spring-Boot): Upgrade from 2.3.2 to 2.4.13
* build(Spring-Boot): Upgrade to 2.4.13 #70

- Required upgrading jedis 3.1.0->3.7.0 and restarting Redis
- Required breaking websocket endpoints into their own controllers

* build(Spring-Boot): Upgrade to 2.4.13 #70

- Fixed issue where accessing admin path was forbidden

* build(Spring-Boot): Upgrade to 2.4.13 #70

Fixed issue where admin who impersonated user was not redirected to the
new user's homepage

* feat(Websockets): Make teacher endpoints secure

#70
