+++ pom.xml
@@ -330,7 +330,7 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.3.2.RELEASE</version>
+    <version>2.4.13</version>
     <relativePath/>
   </parent>
   <dependencies>
@@ -509,17 +509,10 @@
       <artifactId>lombok</artifactId>
       <scope>provided</scope>
     </dependency>
-    <!--
-    <dependency>
-      <groupId>net.bull.javamelody</groupId>
-      <artifactId>javamelody-spring-boot-starter</artifactId>
-      <version>1.78.0</version>
-    </dependency>
-    -->
     <dependency>
       <groupId>redis.clients</groupId>
       <artifactId>jedis</artifactId>
-      <version>3.1.0</version>
+      <version>3.7.0</version>
     </dependency>
     <dependency>
       <groupId>com.google.api-client</groupId>
