+++ src/main/webapp/portal/javascript/general.js
@@ -1,11 +1,11 @@
-function impersonateUser(username) {
-  $.post( "/api/login/impersonate", {username: username}).done(function( data ) {
-    window.location.href = "/";
+function impersonateUser(username, userType) {
+  $.post( "/api/login/impersonate", {username: username}).always(function( data ) {
+    window.location.href = `/${userType}`;
   });
 }
 
 function switchBackToOriginalUser() {
-$.post( "/api/logout/impersonate").done(function( data ) {
+$.post( "/api/logout/impersonate").always(function( data ) {
     window.location.href = "/teacher";
   });
 }
